Move fake into extension
