+++ src/xcode/ENA/ENA/Source/HTTPClient/Service/Models/Receive/TestResultModel.swift
@@ -5,16 +5,4 @@ struct TestResultModel: Codable {
 	let testResult: Int
 	let sc: Int?
 	let labId: String?
-
-	static func fake(
-		testResult: Int = 0,
-		sc: Int? = nil,
-		labId: String? = nil
-	) -> TestResultModel {
-		TestResultModel(
-			testResult: testResult,
-			sc: sc,
-			labId: labId
-		)
-	}
 }
